requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tiktoken==0.7.0
chromadb==0.5.5
pydantic==2.8.2
beautifulsoup4==4.12.3
chromadb==0.5.5
fastapi==0.111.0
lxml==5.2.2
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
tiktoken==0.7.0
uvicorn[standard]==0.30.0
firecrawl-py
openai
psycopg2-binary
spacy==3.8.7
rapidfuzz
jinja2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib